types: clear the last five full-repo type errors
The release gate I added now runs check_types.py --all, which surfaced five
errors CI's numpy version happens not to produce: two redundant casts, an
ndarray-typed local born from np.zeros, a np.prod result assigned to a float,
and np.random.choice's ndarray passed where a list of preference dicts was
expected (a real mismatch, not just an annotation).

Fixed at the source rather than silenced, so local and CI now agree at zero.

+++ stateset_agents/rewards/multi_objective_reward.py
@@ -147,7 +147,9 @@ class MultiObjectiveRewardFunction(RewardFunction):
         elif self.normalization_method == "geometric_mean":
             # Geometric mean of weighted scores
             if all(s > 0 for s in weighted_scores):
-                final_score = np.prod(weighted_scores) ** (1.0 / len(weighted_scores))
+                final_score = float(
+                    np.prod(weighted_scores) ** (1.0 / len(weighted_scores))
+                )
             else:
                 final_score = 0.0
         else:
